Ajout et correction Algo Genetique + un peu de nettoyage
