+++ Sudoku.Benchmark/Sudoku.Benchmark.csproj
@@ -16,6 +16,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Sudoku.Core\Sudoku.Core.csproj" />
     <ProjectReference Include="..\Sudoku.DancingLink\Sudoku.DancingLink.csproj" />
+    <ProjectReference Include="..\Sudoku.GeneticAlgorithm\Sudoku.GeneticAlgorithm.csproj" />
     <ProjectReference Include="..\Sudoku.Norvig\Sudoku.Norvig.csproj" />
     <ProjectReference Include="..\Sudoku.TechniquesHumaines\Sudoku.TechniquesHumaines.csproj" />
     <ProjectReference Include="..\CNNAlgorithm\Sudoku.CNNAlgorithm.csproj" />
