+++ Sudoku.TechniquesHumaines/BackTrackingState.cs
