+++ Sudoku.sln
@@ -7,20 +7,22 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sudoku.Core", "Sudoku.Core\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sudoku.Benchmark", "Sudoku.Benchmark\Sudoku.Benchmark.csproj", "{89EB5D0A-79B2-46EC-BEF7-FB6EC14DAB8B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.DancingLink", "Sudoku.DancingLink\Sudoku.DancingLink.csproj", "{D2A5403F-9B27-4D7D-A3EF-E9E8FF429C7D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sudoku.DancingLink", "Sudoku.DancingLink\Sudoku.DancingLink.csproj", "{D2A5403F-9B27-4D7D-A3EF-E9E8FF429C7D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.Norvig", "Sudoku.Norvig\Sudoku.Norvig.csproj", "{253A6796-4984-4A24-B83A-9FFB634EB356}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sudoku.Norvig", "Sudoku.Norvig\Sudoku.Norvig.csproj", "{253A6796-4984-4A24-B83A-9FFB634EB356}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.TechniquesHumaines", "Sudoku.TechniquesHumaines\Sudoku.TechniquesHumaines.csproj", "{53AB6EE9-48C4-4D82-8F8E-99A29D255D82}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sudoku.TechniquesHumaines", "Sudoku.TechniquesHumaines\Sudoku.TechniquesHumaines.csproj", "{53AB6EE9-48C4-4D82-8F8E-99A29D255D82}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sudoku.CNNAlgorithm", "CNNAlgorithm\Sudoku.CNNAlgorithm.csproj", "{4165EB37-FADD-41FC-84EA-FC3055E5E528}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Suduku.RecursiveSearch", "Suduku.RecursiveSearch\Suduku.RecursiveSearch.csproj", "{5AE6D36E-64DD-48A8-ACE6-36FEF37C4F5E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Suduku.RecursiveSearch", "Suduku.RecursiveSearch\Suduku.RecursiveSearch.csproj", "{5AE6D36E-64DD-48A8-ACE6-36FEF37C4F5E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.CSP", "Sudoku.CSP\Sudoku.CSP.csproj", "{CDEC2146-A020-4795-B78C-A1D72FE35DD1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sudoku.CSP", "Sudoku.CSP\Sudoku.CSP.csproj", "{CDEC2146-A020-4795-B78C-A1D72FE35DD1}"
 	ProjectSection(ProjectDependencies) = postProject
 		{7A604142-5BE0-4C9F-A61C-86B09D8F83E4} = {7A604142-5BE0-4C9F-A61C-86B09D8F83E4}
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.GeneticAlgorithm", "Sudoku.GeneticAlgorithm\Sudoku.GeneticAlgorithm.csproj", "{B530AFD7-605E-485C-829E-AD939CE054B5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -60,6 +62,10 @@ Global
 		{CDEC2146-A020-4795-B78C-A1D72FE35DD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CDEC2146-A020-4795-B78C-A1D72FE35DD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CDEC2146-A020-4795-B78C-A1D72FE35DD1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B530AFD7-605E-485C-829E-AD939CE054B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B530AFD7-605E-485C-829E-AD939CE054B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B530AFD7-605E-485C-829E-AD939CE054B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B530AFD7-605E-485C-829E-AD939CE054B5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
